requirements.txt
# Use Python 3.10.4
# If using conda, name it chatbot3
flask==3.0.3
flask-cors==4.0.1
langchain==0.2.12
langchain_community==0.2.11
langchain_openai==0.1.20
langchain_pinecone==0.1.3
numpy==1.26.4
pinecone-client==5.0.1
python-dotenv==1.0.1
requests==2.32.3
scikit_learn==1.5.1
tiktoken==0.7.0